c4av_model/requirements.txt
pytorch=1.4.0
torchvision=0.5.0
numpy=1.18.1
spacy=2.2.4
pillow=7.0.0
matplotlib=3.1.3
